requirements.txt
langchain_community
openai
huggingface_hub
python-dotenv
ipykernel
beautifulsoup4
streamlit